requirements.txt
Flask>=2.0
Flask-Cors>=3.0
python-dotenv>=0.19
openai-whisper>=1.0
Werkzeug>=2.0

